Added Merge Sort method

+++ AlgDataAPI/Algorithms.API/DTOs/SortResultDTO.cs
@@ -3,7 +3,7 @@
     public class SortResultDTO
     {
         public List<int> SortedValues { get; set; }
-        public float TimeOfCalculation { get; set; }
+        public double TimeOfCalculation { get; set; }
         public int CountOfComparisons { get; set; }
 
     }
